ai-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.9
numpy==1.26.4
python-dotenv==1.0.1
pgvector==0.2.5
requests==2.31.0
huggingface_hub>=0.35.0
faiss-cpu==1.8.0
redis==5.0.4
openai